requirements.txt
Flask>=3.0.0
PyPDF2>=3.0.0
python-docx>=1.0.0
